+++ cppnet/dispatcher.h
@@ -45,7 +45,7 @@ public:
 
     uint32_t AddTimer(const user_timer_call_back& cb, void* param, uint32_t interval, bool always = false);
     uint32_t AddTimer(std::shared_ptr<RWSocket> sock, uint32_t interval, bool always = false);
-    void StopTimer(uint64_t timer_id);
+    void StopTimer(uint32_t timer_id);
 
     std::thread::id GetThreadID() { return _local_thread_id; }
